embedding-updater/requirements.txt
tensorflow==2.13.0
huggingface_hub
Pillow